// src/app/interfaces/mazo.interface.ts
// Generated by https://quicktype.io

export interface Mazo {
  success:   boolean;
  deck_id:   string;
  cards:     Card[];
  remaining: number;
}

export interface Card {
  code:   string;
  image:  string;
  images: Images;
  value:  string;
  suit:   string;
  number: number;
}

export interface Images {
  svg: string;
  png: string;
}